refactor : 패키지 변경 및 스키마 변경
 - ordertable -> order
 - order의 fk not null -> nullable
