+++ src/main/java/kitchenpos/order/application/OrderTableService.java
@@ -1,25 +1,27 @@
-package kitchenpos.ordertable.application;
+package kitchenpos.order.application;
 
-import kitchenpos.ordertable.domain.ChangeService;
-import kitchenpos.ordertable.domain.OrderTable;
-import kitchenpos.ordertable.dto.OrderTableCreateRequest;
-import kitchenpos.ordertable.dto.OrderTableEmptyChangeRequest;
-import kitchenpos.ordertable.dto.OrderTableGuestsChangeRequest;
-import kitchenpos.ordertable.dto.OrderTableResponse;
-import kitchenpos.ordertable.repository.OrderTableRepository;
+import kitchenpos.order.domain.OrderStatus;
+import kitchenpos.order.domain.OrderTable;
+import kitchenpos.order.dto.request.OrderTableCreateRequest;
+import kitchenpos.order.dto.request.OrderTableEmptyChangeRequest;
+import kitchenpos.order.dto.request.OrderTableGuestsChangeRequest;
+import kitchenpos.order.dto.response.OrderTableResponse;
+import kitchenpos.order.repository.OrderRepository;
+import kitchenpos.order.repository.OrderTableRepository;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
+import java.util.Arrays;
 import java.util.List;
 
 @Service
 public class OrderTableService {
     private final OrderTableRepository orderTableRepository;
-    private final ChangeService changeService;
+    private final OrderRepository orderRepository;
 
-    public OrderTableService(OrderTableRepository orderTableRepository, ChangeService changeService) {
+    public OrderTableService(OrderTableRepository orderTableRepository, OrderRepository orderRepository) {
         this.orderTableRepository = orderTableRepository;
-        this.changeService = changeService;
+        this.orderRepository = orderRepository;
     }
 
     @Transactional
@@ -38,7 +40,7 @@ public class OrderTableService {
         OrderTable savedOrderTable = orderTableRepository.findById(orderTableId)
                 .orElseThrow(IllegalArgumentException::new);
 
-        boolean isCookingOrMeal = changeService.isCookingOrMeal(orderTableId);
+        boolean isCookingOrMeal = orderRepository.existsByOrderTableIdAndOrderStatusIn(orderTableId, Arrays.asList(OrderStatus.COOKING, OrderStatus.MEAL));
 
         savedOrderTable.changeEmpty(request.isEmpty(), isCookingOrMeal);
