+++ src/main/java/kitchenpos/order/dto/request/OrderTableCreateRequest.java
@@ -1,4 +1,4 @@
-package kitchenpos.ordertable.dto;
+package kitchenpos.order.dto.request;
 
 import javax.validation.constraints.NotNull;
