+++ src/main/java/kitchenpos/order/repository/TableGroupRepository.java
@@ -1,6 +1,6 @@
-package kitchenpos.ordertable.repository;
+package kitchenpos.order.repository;
 
-import kitchenpos.ordertable.domain.TableGroup;
+import kitchenpos.order.domain.TableGroup;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface TableGroupRepository extends JpaRepository<TableGroup, Long> {
