--- src/main/java/kitchenpos/ordertable/domain/ChangeService.java
@@ -1,7 +1,0 @@
-package kitchenpos.ordertable.domain;
-
-public interface ChangeService {
-    void ungroup(OrderTables orderTables);
-
-    boolean isCookingOrMeal(Long orderTableId);
-}
