+++ src/main/resources/db/migration/V1__Initialize_project_tables.sql
@@ -1,8 +1,9 @@
-CREATE TABLE orders (
-    id BIGINT(20) NOT NULL AUTO_INCREMENT,
-    order_table_id BIGINT(20) NOT NULL,
-    order_status VARCHAR(255) NOT NULL,
-    ordered_time DATETIME NOT NULL,
+CREATE TABLE orders
+(
+    id             BIGINT(20)   NOT NULL AUTO_INCREMENT,
+    order_table_id BIGINT(20),
+    order_status   VARCHAR(255) NOT NULL,
+    ordered_time   DATETIME     NOT NULL,
     PRIMARY KEY (id)
 );
