+++ src/test/java/kitchenpos/ordertable/domain/OrderTablesTest.java
@@ -2,7 +2,10 @@ package kitchenpos.ordertable.domain;
 
 import kitchenpos.order.domain.Order;
 import kitchenpos.order.domain.OrderStatus;
+import kitchenpos.order.domain.OrderTable;
+import kitchenpos.order.domain.OrderTables;
 import kitchenpos.order.domain.Orders;
+import kitchenpos.order.domain.TableGroup;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
